Change Group/XIA IDs from u64 to u32 and remove all as-any casts in WasmBridge
Root cause: wasm-bindgen converts Rust u64 to JS bigint, but TypeScript
uses number — forcing 12 as-any casts in WasmBridge.ts. Changing IDs to
u32 (sufficient for 4B+ entities) makes wasm-bindgen emit number directly.

Rust changes:
- xia.rs: XiaId = u32
- group.rs: GroupId/ComponentDefId/ComponentInstanceId = u32
- scene.rs: next_xia_id u32 (snapshot format kept u64 for backward compat)
- lib.rs: all WASM group/xia params u64→u32, get_xia_for_face returns u32

TypeScript changes:
- WasmBridge.ts: 12 as-any → 0 (complete elimination)
- AxiaEngineExtended: added getPositionsF64, getDirtyFaceBuffers,
  is_face_locked, get_xia_for_face with proper WasmDeltaBuffers interface
- No bigint anywhere in .d.ts

+++ crates/axia-core/src/group.rs
@@ -29,13 +29,13 @@ use axia_geo::FaceId;
 // ════════════════════════════════════════════════════════════════
 
 /// 그룹 고유 식별자
-pub type GroupId = u64;
+pub type GroupId = u32;
 
 /// 컴포넌트 정의 고유 식별자
-pub type ComponentDefId = u64;
+pub type ComponentDefId = u32;
 
 /// 컴포넌트 인스턴스 고유 식별자
-pub type ComponentInstanceId = u64;
+pub type ComponentInstanceId = u32;
 
 // ════════════════════════════════════════════════════════════════
 // Transform3D
@@ -213,9 +213,9 @@ pub struct GroupManager {
     pub component_defs: HashMap<ComponentDefId, ComponentDef>,
     pub component_instances: HashMap<ComponentInstanceId, ComponentInstance>,
 
-    next_group_id: u64,
-    next_def_id: u64,
-    next_instance_id: u64,
+    next_group_id: u32,
+    next_def_id: u32,
+    next_instance_id: u32,
 
     /// face → group 역인덱스 (빠른 조회)
     #[serde(skip)]
@@ -488,7 +488,7 @@ impl GroupManager {
     pub fn export_group_info(&self, group_id: GroupId) -> Option<String> {
         let group = self.groups.get(&group_id)?;
         let face_ids: Vec<u32> = group.face_ids.iter().map(|f| f.raw()).collect();
-        let children: Vec<u64> = group.children.clone();
+        let children: Vec<u32> = group.children.clone();
 
         Some(format!(
             r#"{{"id":{},"name":"{}","faceIds":{:?},"children":{:?},"visible":{},"locked":{},"parent":{},"isComponent":{}}}"#,
